requirements.txt
Django==4.0.4
django-environ==0.8.1
djangorestframework==3.13.1
psycopg2==2.9.3
django-crispy-forms==1.14.0
drf-spectacular==0.22.0
requests==2.27.1
django-filter==21.1
